+++ devctl/utils/__init__.py
@@ -1,3 +1,5 @@
 """
 General utility functions and helpers.
 """
+
+from devctl.utils.platform import get_platform as get_platform